requirements.txt
alibabacloud_dysmsapi20170525==2.0.24
alibabacloud_green20220302==1.0.5
alibabacloud_tea_openapi==0.3.7
alipay-sdk-python==3.6.703
argon2-cffi==21.3.0
attrs==23.1.0
channels==4.0.0
daphne==4.0.0
Django==4.2.3
django-cors-headers==4.2.0
djangorestframework==3.14.0
jieba==0.42.1
langchain==0.0.240
openai==0.27.4
protobuf==4.23.4
protobuf_inspector==0.2
python-dateutil==2.8.2
python_docx==0.8.11
levenshtein==0.21.1
requests==2.31.0
websocket-client==1.6.1
